+++ R/list_channel_resources.R
@@ -4,6 +4,7 @@
 #' @param part a comma separated list of channel resource properties that response will include
 #' string, required, one of the following: auditDetails, brandingSettings, contentDetails, contentOwnerDetails, id, invideoPromotion, localizations, snippet, statistics, status, topicDetails
 #' default is contentDetails
+#' @param hl  language that will be used for text values, optional, default is en-US. See also \code{\link{list_langs}}
 #' @param max_results maximum number of items that should be returned, integer, can be between 0 and 50, default is 50
 #' @param page_token specific page in the result set that should be returned, optional
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.