+++ R/list_guidecats.R
@@ -1,6 +1,7 @@
 #' Get List of categories that can be associated with YouTube channels
 #' 
-#' @param regionCode Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search}).
+#' @param regionCode Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search})
+#' @param hl  language that will be used for text values, optional, default is en-US. See also \code{\link{list_langs}}
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @return data.frame with 4 columns: channelId, title, etag, id
@@ -11,11 +12,11 @@
 #' list_guidecats("JP")
 #' }
 
-list_guidecats <- function (regionCode=NULL, ...) {
+list_guidecats <- function (regionCode=NULL, hl = NULL, ...) {
 
 	if (is.null(regionCode)) stop("Must specify a regionCode")
 
-	querylist <- list(part="snippet", regionCode=regionCode)
+	querylist <- list(part="snippet", regionCode=regionCode, hl = NULL)
 
 	res <- tuber_GET("guideCategories", querylist, ...)